+++ packages/app/components/landing_page/LandingPageAccordion.tsx
@@ -5,6 +5,8 @@ import { MaterialIcons } from '@expo/vector-icons';
 import { theme } from '../../theme';
 import { useState } from 'react';
 import useCustomStyles from 'app/hooks/useCustomStyles';
+import useAccordionState from './useAccordionState';
+import loadStyles from './landingPageStyles';
 
 export const LandingPageAccordion = ({ title, content, iconName }) => {
   const styles = useCustomStyles(loadStyles);
@@ -35,48 +37,4 @@ export const LandingPageAccordion = ({ title, content, iconName }) => {
       )}
     </RCard>
   );
-};
-
-const loadStyles = (theme) => {
-  const { currentTheme } = theme;
-  return {
-    card: {
-      marginBottom: 10,
-      width: '100%',
-      backgroundColor: currentTheme.colors.secondaryBlue,
-    },
-    cardHeader: {
-      flexDirection: 'row',
-      alignItems: 'center',
-      justifyContent: 'space-between',
-      paddingHorizontal: 20,
-      paddingVertical: 10,
-    },
-    transparentButton: {
-      backgroundColor: 'transparent',
-    },
-    icon: {
-      fontSize: 28,
-      color: '#34a89a',
-      marginRight: 10,
-    },
-    featureText: {
-      fontSize: 18,
-      color: currentTheme.colors.text,
-    },
-    cardContent: {
-      paddingHorizontal: 20,
-      paddingVertical: 10,
-      fontSize: 16,
-      color: currentTheme.colors.text,
-    },
-  };
-};
-
-const useAccordionState = () => {
-  const [expanded, setExpanded] = useState(false);
-
-  const toggleExpanded = () => setExpanded((prevState) => !prevState);
-
-  return [expanded, toggleExpanded];
 };
