+++ packages/app/components/landing_page/useAccordionState.tsx
@@ -1,0 +1,10 @@
+import { useState } from 'react';
+const useAccordionState = () => {
+  const [expanded, setExpanded] = useState(false);
+
+  const toggleExpanded = () => setExpanded((prevState) => !prevState);
+
+  return [expanded, toggleExpanded];
+};
+
+export default useAccordionState;
